# backend/swagger.json
{"openapi":"3.0.0","paths":{"/auth/me":{"get":{"operationId":"AuthController_me","parameters":[],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"Return the application principal for the OIDC subject","tags":["authentication"]}},"/todos":{"post":{"operationId":"TodoController_addTodo","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AddTodoRequestDto"}}}},"responses":{"201":{"description":"Todo created successfully"},"400":{"description":"Invalid input"}},"security":[{"bearer":[]}],"summary":"Create a new todo","tags":["todos"]},"get":{"operationId":"TodoController_getAll","parameters":[{"name":"limit","required":true,"in":"query","schema":{"type":"number"}},{"name":"offset","required":true,"in":"query","schema":{"type":"number"}}],"responses":{"200":{"description":"Returns all todos","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetAllTodosResponseDto"}}}}},"security":[{"bearer":[]}],"summary":"Get all todos","tags":["todos"]}},"/todos/{id}/complete":{"patch":{"operationId":"TodoController_completeTodo","parameters":[{"name":"id","required":true,"in":"path","description":"Todo ID","schema":{"type":"string"}}],"responses":{"204":{"description":"Todo marked as completed"},"400":{"description":"Invalid todo ID"},"404":{"description":"Todo not found"}},"security":[{"bearer":[]}],"summary":"Mark a todo as completed","tags":["todos"]}},"/todos/{id}/uncomplete":{"patch":{"operationId":"TodoController_uncompleteTodo","parameters":[{"name":"id","required":true,"in":"path","description":"Todo ID","schema":{"type":"string"}}],"responses":{"204":{"description":"Todo marked as uncompleted"},"400":{"description":"Invalid todo ID"},"404":{"description":"Todo not found"}},"security":[{"bearer":[]}],"summary":"Mark a todo as uncompleted","tags":["todos"]}},"/todos/{id}/title":{"patch":{"operationId":"TodoController_modifyTitle","parameters":[{"name":"id","required":true,"in":"path","description":"Todo ID","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ModifyTodoTitleRequestDto"}}}},"responses":{"204":{"description":"Todo title updated"},"400":{"description":"Invalid input"},"404":{"description":"Todo not found"}},"security":[{"bearer":[]}],"summary":"Update a todo title","tags":["todos"]}},"/todos/{id}":{"delete":{"operationId":"TodoController_deleteTodo","parameters":[{"name":"id","required":true,"in":"path","description":"Todo ID","schema":{"type":"string"}}],"responses":{"204":{"description":"Todo deleted"},"400":{"description":"Invalid todo ID"},"404":{"description":"Todo not found"}},"security":[{"bearer":[]}],"summary":"Delete a todo","tags":["todos"]}},"/sse/todos/stream":{"get":{"operationId":"TodoSSEController_stream","parameters":[],"responses":{"200":{"description":""}},"tags":["TodoSSE"]}},"/sse/todos/subscriptions":{"post":{"operationId":"TodoSSEController_subscribe","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TodoSubscriptionDto"}}}},"responses":{"201":{"description":""}},"tags":["TodoSSE"]}}},"info":{"title":"API Documentation","description":"API description","version":"1.0.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"AddTodoRequestDto":{"type":"object","properties":{"title":{"type":"string","description":"The title of the todo"}},"required":["title"]},"TodoDto":{"type":"object","properties":{"id":{"type":"string","description":"The unique identifier of the todo"},"title":{"type":"string","description":"The title of the todo"},"completed":{"type":"boolean","description":"The completion status of the todo"},"createdAt":{"type":"number","description":"The creation timestamp of the todo"},"updatedAt":{"type":"number","description":"The last update timestamp of the todo"}},"required":["id","title","completed","createdAt"]},"GetAllTodosResponseDto":{"type":"object","properties":{"todos":{"description":"Array of todo items","type":"array","items":{"$ref":"#/components/schemas/TodoDto"}}},"required":["todos"]},"ModifyTodoTitleRequestDto":{"type":"object","properties":{"id":{"type":"string","description":"The unique identifier of the todo"},"title":{"type":"string","description":"The new title of the todo"}},"required":["id","title"]},"TodoSubscriptionDto":{"type":"object","properties":{"subscriberId":{"type":"string","example":"tab-7f3a2c"},"events":{"example":["todo.added","todo.completed"],"type":"array","items":{"type":"string"}}},"required":["subscriberId","events"]}}}}